Refactor headers in panda/board/jungle/boards/.*h
